change package name in cdi
