+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.1/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-io.astefanutti.metrics21.cdi.MetricsExtension21
+com.ibm.ws.microprofile.metrics21.cdi.MetricsExtension21
