+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.1/src/com/ibm/ws/microprofile/metrics21/cdi/MetricsInterceptor21.java
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package io.astefanutti.metrics21.cdi;
+package com.ibm.ws.microprofile.metrics21.cdi;
 
 import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.InvocationTargetException;
