+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.1/src/com/ibm/ws/microprofile/metrics21/cdi/package-info.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2017, 2019 IBM Corporation and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -26,6 +26,6 @@
  */
 @org.osgi.annotation.versioning.Version("1.0")
 @TraceOptions(traceGroup = "METRICS", messageBundle = "com.ibm.ws.microprofile.metrics.cdi.resources.MetricsCDI")
-package io.astefanutti.metrics21.cdi;
+package com.ibm.ws.microprofile.metrics21.cdi;
 
 import com.ibm.websphere.ras.annotation.TraceOptions;
